// src/domain/interfaces/repositories/_gen/ActionPlanReportResultRepository.ts
// This file is generated by scripts/generateInterfaceRepositories.ts
import { ActionPlan } from '../../../entities/ActionPlan'
import { ActionPlanReportResult } from '../../../entities/ActionPlanReportResult'
import { PromisedResult, UnknownRuntimeError, AlreadyExistsError, NotFoundError } from '../../../../_shared/error'

export interface ActionPlanReportResultRepository {
  issueId: () => PromisedResult<string, UnknownRuntimeError>
  getAll: () => PromisedResult<ActionPlanReportResult[], UnknownRuntimeError>
  getRelevant: (text: string, limit: number) => PromisedResult<ActionPlanReportResult[], UnknownRuntimeError>
  create: (entity: ActionPlanReportResult) => PromisedResult<ActionPlanReportResult, UnknownRuntimeError | AlreadyExistsError>
  update: (entity: ActionPlanReportResult) => PromisedResult<ActionPlanReportResult, UnknownRuntimeError | NotFoundError>
  delete: (id: string) => PromisedResult<void, UnknownRuntimeError | NotFoundError>
  getByActionPlanId: (ActionPlanId: ActionPlan['id']) => PromisedResult<ActionPlanReportResult | null, UnknownRuntimeError>
}