// src/domain/interfaces/repositories/_gen/ActionResultCollectDataRepository.ts
// This file is generated by scripts/generateInterfaceRepositories.ts
import { ActionResult } from '../../../entities/ActionResult'
import { ActionResultCollectData } from '../../../entities/ActionResultCollectData'
import { PromisedResult, UnknownRuntimeError, AlreadyExistsError, NotFoundError } from '../../../../_shared/error'

export interface ActionResultCollectDataRepository {
  issueId: () => PromisedResult<string, UnknownRuntimeError>
  getAll: () => PromisedResult<ActionResultCollectData[], UnknownRuntimeError>
  getRelevant: (text: string, limit: number) => PromisedResult<ActionResultCollectData[], UnknownRuntimeError>
  create: (entity: ActionResultCollectData) => PromisedResult<ActionResultCollectData, UnknownRuntimeError | AlreadyExistsError>
  update: (entity: ActionResultCollectData) => PromisedResult<ActionResultCollectData, UnknownRuntimeError | NotFoundError>
  delete: (id: string) => PromisedResult<void, UnknownRuntimeError | NotFoundError>
  getByActionResultId: (ActionResultId: ActionResult['id']) => PromisedResult<ActionResultCollectData | null, UnknownRuntimeError>
}